refactor: 기획자/개발자 각 기능별 섹션 분리
